# docs/swagger.yaml
definitions:
...
  ds.Chat:
    properties:
      friends:
        type: integer
      id:
        type: integer
      img:
        type: string
      info:
        type: string
      isDelete:
        type: boolean
      name:
        type: string
      nickname:
        type: string
      subscribers:
        type: integer
    type: object
  ds.ChatRequest:
    properties:
      friends:
        type: integer
      info:
        type: string
      name:
        type: string
      nickname:
        type: string
      subscribers:
        type: integer
    type: object
  ds.ChatResponse:
    properties:
      friends:
        type: integer
      id:
        type: integer
      img:
        type: string
      info:
        type: string
      name:
        type: string
      nickname:
        type: string
      subscribers:
        type: integer
    type: object
...
  ds.ErrorResp:
    properties:
      description:
        type: string
      status:
        type: string
    type: object
  ds.GetChatsResponse:
    properties:
      chats:
        items:
          $ref: '#/definitions/ds.ChatResponse'
        type: array
      draft_ID:
        type: integer
      draft_count:
        type: integer
    type: object
...
  ds.OkResp:
    properties:
      message:
        type: string
    type: object
...
info:
  contact: {}
paths:
  /chat/{id}/in-message:
    post:
      description: Добавляет чат к конкретному сообщению.
      parameters:
      - description: ID чата
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Сообщение о добавлении чата
          schema:
            $ref: '#/definitions/ds.OkResp'
        "400":
          description: Некорректный ID чата
          schema:
            $ref: '#/definitions/ds.ErrorResp'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/ds.ErrorResp'
      security:
      - BearerAuth: []
      summary: Добавить чат в сообщение
      tags:
      - Chats
  /chats:
    get:
      description: Возвращает список чатов для конкретного пользователя с указанием
        черновиков.
      parameters:
      - description: Фильтр по имени чата
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Список чатов с черновиками
          schema:
            $ref: '#/definitions/ds.GetChatsResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/ds.ErrorResp'
      tags:
      - Chats
  /chats/{id}:
    delete:
      description: Удаляет чат по его ID и удаляет изображение из Minio.
      parameters:
      - description: ID чата
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Сообщение об успешном удалении
          schema:
            $ref: '#/definitions/ds.OkResp'
        "400":
          description: Некорректный ID чата
          schema:
            $ref: '#/definitions/ds.ErrorResp'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/ds.ErrorResp'
      security:
      - BearerAuth: []
      summary: Удалить чат
      tags:
      - Chats
    get:
      description: Возвращает информацию о чате по его ID.
      parameters:
      - description: ID чата
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Информация о чате
          schema:
            $ref: '#/definitions/ds.ChatResponse'
        "400":
          description: Некорректный ID чата
          schema:
            $ref: '#/definitions/ds.ErrorResp'
        "404":
          description: Чат не найден
          schema:
            $ref: '#/definitions/ds.ErrorResp'
      summary: Получить чат по ID
      tags:
      - Chats
    put:
      consumes:
      - application/json
      description: Обновляет данные существующего чата.
      parameters:
      - description: ID чата
        in: path
        name: id
        required: true
        type: integer
      - description: Обновленные данные чата
        in: body
        name: chat
        required: true
        schema:
          $ref: '#/definitions/ds.ChatRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Обновлённый чат
          schema:
            $ref: '#/definitions/ds.Chat'
        "400":
          description: Некорректный ID чата или неверные данные
          schema:
            $ref: '#/definitions/ds.ErrorResp'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/ds.ErrorResp'
      security:
      - BearerAuth: []
      summary: Обновить чат
      tags:
      - Chats
  /chats/{id}/new-image:
    post:
      consumes:
      - multipart/form-data
      description: Загружает и заменяет изображение чата.
      parameters:
      - description: ID чата
        in: path
        name: id
        required: true
        type: integer
      - description: Новое изображение для чата
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Сообщение об успешной загрузке
          schema:
            $ref: '#/definitions/ds.OkResp'
        "400":
          description: Невалидные данные или неправильный формат файла
          schema:
            $ref: '#/definitions/ds.ErrorResp'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/ds.ErrorResp'
      security:
      - BearerAuth: []
      summary: Заменить изображение чата
      tags:
      - Chats
  /chats/create:
    post:
      consumes:
      - application/json
      description: Создает новый чат с указанными данными.
      parameters:
      - description: Информация о чате
        in: body
        name: chat
        required: true
        schema:
          $ref: '#/definitions/ds.ChatRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Созданный чат
          schema:
            $ref: '#/definitions/ds.Chat'
        "400":
          description: Неверные данные или отсутствует имя/никнейм
          schema:
            $ref: '#/definitions/ds.ErrorResp'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/ds.ErrorResp'
      security:
      - BearerAuth: []
      summary: Создать новый чат
      tags:
      - Chats
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"